Moved check-signature to new Next structure
